+++ src/idmatcherai/matcher.py
@@ -1,17 +1,26 @@
+import logging
 from pathlib import Path
 
 import numpy as np
-from scipy.optimize import linear_sum_assignment as hungarian
-from scipy.sparse import coo_matrix
+from scipy.optimize import linear_sum_assignment
 
 from idtrackerai import Fragment
 from idtrackerai.tracker.network.get_predictions import GetPredictionsIdentities
 
+from .images import extact_all_images_and_labels
 from .network import load_identification_model
 
 
-def match(model_folder: Path, images: np.ndarray, labels: np.ndarray):
-    identification_model, model_params = load_identification_model(model_folder)
+def match(id_images_path: Path, model_path: Path):
+    logging.info(
+        "Matching images from %s with model from %s", id_images_path, model_path
+    )
+
+    images, labels = extact_all_images_and_labels(
+        id_images_path.glob("id_images_*.hdf5")
+    )
+    model, model_params = load_identification_model(model_path)
+
     set_of_labels = set(labels.astype(int))
     set_of_labels.discard(0)
 
@@ -22,17 +31,17 @@ def match(model_folder: Path, images: np.ndarray, labels: np.ndarray):
     # TODO if num_labels <= num_classes:
 
     confusion_matrix = np.zeros((num_classes, num_labels))
-    frequencies_matrix = np.zeros((num_classes, num_labels))
+    frequencies_matrix = np.zeros((num_classes, num_labels), int)
     certainties = np.zeros(num_labels)
 
     for identity in set_of_labels:
         assigner = GetPredictionsIdentities(
-            identification_model, images[labels == identity], model_params
+            model, images[labels == identity], model_params
         )
         assigner.get_all_predictions()
         frequencies, P1_vector, certainty = compute_identification_statistics(assigner)
-        confusion_matrix[:, identity - 1] = P1_vector
-        frequencies_matrix[:, identity - 1] = frequencies
+        confusion_matrix[identity - 1] = P1_vector
+        frequencies_matrix[identity - 1] = frequencies
         certainties[identity - 1] = certainty
     return confusion_matrix, frequencies_matrix, certainties
 
@@ -43,66 +52,17 @@ def get_transfer_dicts(confusion_matrix: np.ndarray, frequencies_matrix: np.ndar
     :param frequencies_matrix: rows are identities in model, columns are identities in images
     :return: transfer_dicts: in each dictionary: the key is the identity of the image, the value is the identity assigned by the model
     """
-    num_classes = confusion_matrix.shape[0]
-    normalized_frequencies = 1.0 - frequencies_matrix / np.tile(
-        np.sum(frequencies_matrix, axis=0), [num_classes, 1]
-    )
     return {
-        "max_P1": get_trasnfer_dict_by_maximum(confusion_matrix),
-        "max_freq": get_trasnfer_dict_by_maximum(frequencies_matrix),
-        "greedy": get_transfer_dict_greedy(confusion_matrix.copy()),
-        "hungarian_P1": get_transfer_dict_by_hungarian(1.0 - confusion_matrix),
-        "hungarian_freq": get_transfer_dict_by_hungarian(normalized_frequencies),
+        "hungarian_P1": get_transfer_dict_by_hungarian(confusion_matrix),
+        "hungarian_freq": get_transfer_dict_by_hungarian(
+            frequencies_matrix / np.sum(frequencies_matrix, axis=1)[:, None]
+        ),
     }
 
 
 def get_transfer_dict_by_hungarian(matrix):
-    num_labels = matrix.shape[1]
-    r, assign = hungarian(matrix)
-    transfer_dict = {
-        "assignments_values": {i: None for i in range(1, num_labels + 1)},
-        "assignments": {i: None for i in range(1, num_labels + 1)},
-    }
-    for rr, a in zip(r, assign):
-        transfer_dict["assignments"][a + 1] = rr + 1
-        transfer_dict["assignments_values"][a + 1] = matrix[rr, a]
-
-    return transfer_dict
-
-
-def get_trasnfer_dict_by_maximum(matrix):
-    num_labels = matrix.shape[1]
-    transfer_dict = {
-        "assignments_values": {i: None for i in range(1, num_labels + 1)},
-        "assignments": {i: None for i in range(1, num_labels + 1)},
-    }
-    for identity in range(1, num_labels + 1):
-        transfer_dict["assignments"][identity] = np.argmax(matrix[:, identity - 1]) + 1
-        transfer_dict["assignments_values"][identity] = np.max(matrix[:, identity - 1])
-
-    return transfer_dict
-
-
-def get_transfer_dict_greedy(confusion_matrix):
-    num_labels = confusion_matrix.shape[1]
-    transfer_dict = {
-        "assignments_values": {i: None for i in range(1, num_labels + 1)},
-        "assignments": {i: None for i in range(1, num_labels + 1)},
-    }
-    max_assigned = np.max(confusion_matrix, axis=0)
-    order_of_assignment = np.argsort(max_assigned)[::-1]  # 0 indexed
-
-    for id_to_assign in order_of_assignment:
-        id_assigned = np.argmax(confusion_matrix[:, id_to_assign]) + 1
-        transfer_dict["assignments"][id_to_assign + 1] = id_assigned
-        transfer_dict["assignments_values"][id_to_assign + 1] = np.max(
-            confusion_matrix[:, id_to_assign]
-        )
-
-        confusion_matrix[id_assigned - 1, :] = 0
-        confusion_matrix[:, id_to_assign] = 0
-
-    return transfer_dict
+    r, assign = linear_sum_assignment(matrix, maximize=True)
+    return {"assignments_values": matrix[r, assign], "assignments": r + 1}
 
 
 def compute_identification_statistics(assigner: GetPredictionsIdentities):
@@ -180,42 +140,26 @@ def compute_unique_matches_by_confusion_matrix(confusion_matrix, threshold=0.99)
 
 
 def check_bidirectional_matches(
-    transfer_dicts_A_B, transfer_dicts_B_A, num_animals_A, num_animals_B
+    transfer_dicts_A_B: dict,
+    transfer_dicts_B_A: dict,
+    n_animals_A: int,
+    n_animals_B: int,
 ):
     matches_dict = {}
     for assignment_type in transfer_dicts_A_B.keys():
-        matches_dict[assignment_type] = {}
+        matrix_A_B = np.zeros((n_animals_A, n_animals_B), bool)
         assignments = transfer_dicts_A_B[assignment_type]["assignments"]
-        rows = (
-            np.asarray([av for av in assignments.values() if av is not None]) - 1
-        )  # identities in model A
-        cols = (
-            np.asarray([ak for ak in assignments.keys() if assignments[ak] is not None])
-            - 1
-        )  # identities in images B
-        assert len(rows) == len(cols)
-        matrix_A_B = coo_matrix(
-            (np.ones(len(rows)), (rows, cols)), shape=(num_animals_A, num_animals_B)
-        ).toarray()
+        matrix_A_B[range(len(assignments)), assignments - 1] = True
 
+        matrix_B_A = np.zeros((n_animals_B, n_animals_A), bool)
         assignments = transfer_dicts_B_A[assignment_type]["assignments"]
-        rows = np.asarray([av for av in assignments.values() if av is not None]) - 1
-        cols = (
-            np.asarray([ak for ak in assignments.keys() if assignments[ak] is not None])
-            - 1
-        )
-        assert len(rows) == len(cols)
-        matrix_B_A = coo_matrix(
-            (np.ones(len(rows)), (rows, cols)), shape=(num_animals_B, num_animals_A)
-        ).toarray()
+        matrix_B_A[range(len(assignments)), assignments - 1] = True
 
-        matches = np.where(matrix_A_B * matrix_B_A.T == 1)
-        matches_dict[assignment_type]["unique_matches"] = list(
-            zip(matches[0] + 1, matches[1] + 1)
-        )
-        missmatches = np.where(matrix_A_B != matrix_B_A.T)
-        matches_dict[assignment_type]["missmatches"] = list(
-            zip(missmatches[0] + 1, missmatches[1] + 1)
-        )
+        matches = np.where(matrix_A_B * matrix_B_A.T)
+        missmatches = np.where(np.logical_xor(matrix_A_B, matrix_B_A.T))
+        matches_dict[assignment_type] = {
+            "unique_matches": list(zip(matches[0] + 1, matches[1] + 1)),
+            "missmatches": list(zip(missmatches[0] + 1, missmatches[1] + 1)),
+        }
 
     return matches_dict